+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 unreleased
 ==========
+* fix: Content Expiry changelist filter dates should be for the future not the past
 
 0.0.1 (2021-10-04)
 ==================